Adding init scripts
