+++ config/settings.py
@@ -1,4 +1,5 @@
 # config/settings.py
+import os
 
 BOT_PREFIX = "!"
 DEFAULT_CHANNEL_ID = int(os.getenv("CHANNEL_ID")) # Use 0 or a placeholder if not set
