+++ src/bot/client.py
@@ -4,7 +4,7 @@ from discord.ext import commands
 
 # Import your API clients and core logic classes
 from src.api.sleeper import SleeperClient
-from src.api.rapidapi_nfl import RapidAPINFLClient
+from src.api.rapidapi_nfl import RapidApiNFLClient
 from src.api.ktc_scraper import KTCScraper
 from src.core.player_data_manager import PlayerDataManager
 from src.core.trade_analyzer import TradeAnalyzer # The class for trade wins logic
@@ -12,7 +12,7 @@ from src.db.repositories.trade_repo import TradeRepository # For saving/fetching
 
 class MyDiscordBot(commands.Bot):
     def __init__(self, command_prefix, intents, league_id: str, db_pool, default_channel_id: int, **kwargs):
-        super().__init__(command_prefix=command_prefix, intents=intents, **kwargs)
+        super().__init__(command_prefix=command_prefix, intents=intents, help_command=None, **kwargs)
 
         # Custom properties/attributes accessible across Cogs via bot.sleeper_client etc.
         self.league_id = league_id
@@ -21,7 +21,7 @@ class MyDiscordBot(commands.Bot):
 
         # Initialize API clients (these manage their own connections/sessions)
         self.sleeper_client = SleeperClient(self.league_id)
-        self.rapidapi_nfl_client = RapidAPINFLClient()
+        self.rapidapi_nfl_client = RapidApiNFLClient()
         self.ktc_scraper = KTCScraper()
 
         # Initialize core logic managers
@@ -77,7 +77,11 @@ class MyDiscordBot(commands.Bot):
         
         # Optionally, fetch player data into cache on startup
         print("Warming up player data cache...")
-        await self.player_data_manager.warm_cache()
+        # cache once a day, or on startup
+        # This can be a long operation, so you might want to run it in the background
+        ##########################################################################
+        # await self.player_data_manager.warm_cache()
+        ##########################################################################
         print("Player data cache warmed.")
